requirements.txt
glob2==0.7
PyYAML==6.0.1
youtube-dl==2021.12.17
yt-dlp==2023.7.6
